# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Velo API",
    "version": "0.1.0",
    "description": "Velo — anonymous cash liquidity on Stellar. A lightweight, payment-aware service interface for agent-assisted and application-driven cash flows. Paid routes use an x402-style challenge: call without payment to receive a challenge, pay the quoted USDC amount on Stellar with the `velo:request` memo, then retry with the transaction hash in the `X-Payment` header. Each payment transaction is single-use.",
    "contact": {
      "name": "Nullifier Systems",
      "url": "https://github.com/Nullifier-Systems/velo"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/Nullifier-Systems/velo/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://api.velo.cash",
      "description": "Mainnet (production)"
    }
  ],
...
  "paths": {
...
    "/api/v1/cash/request/{id}": {
      "get": {
        "operationId": "getCashRequest",
        "tags": [
          "cash"
        ],
        "summary": "Poll request status",
        "description": "Free polling endpoint for a pending cash request. The secret is never included.",
        "x-rate-limit": {
          "max": 60,
          "timeWindow": "1 minute"
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Trade id returned when the request was created (64-char hex).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current state of the cash request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashRequestStatus"
                }
              }
            }
          },
          "404": {
            "description": "No cash request with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "x402Payment": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Payment",
        "description": "x402 payment header: the hash of a successful Stellar transaction paying the merchant at least the operation's `x-price-usdc` with memo `velo:request`. Each transaction hash is accepted once. Requests without a valid payment receive 402 with a challenge, not 401."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          },
          "detail": {
            "type": "string",
            "description": "Optional underlying error detail."
          }
        }
      },
...
      "RateLimitError": {
        "type": "object",
        "required": [
          "statusCode",
          "error",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "const": 429
          },
          "error": {
            "type": "string",
            "const": "Too Many Requests"
          },
          "message": {
            "type": "string"
          },
          "retryAfter": {
            "type": "string",
            "description": "Human-readable wait time, e.g. \"1 minute\"."
          },
          "retryAfterSeconds": {
            "type": "integer",
            "description": "Seconds to wait before retrying."
          }
        }
      },
...
      "CashRequestStatus": {
        "type": "object",
        "description": "Public view of a cash request. The redemption secret is never included.",
        "required": [
          "id",
          "contractId",
          "seller",
          "buyer",
          "amountStroops",
          "secretHashHex",
          "qrPayload",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Trade id (64-char hex)."
          },
          "contractId": {
            "type": "string",
            "description": "Escrow contract id (C...)."
          },
          "seller": {
            "type": "string"
          },
          "buyer": {
            "type": "string"
          },
          "amountStroops": {
            "type": "string"
          },
          "secretHashHex": {
            "type": "string"
          },
          "qrPayload": {
            "type": "string",
            "description": "Persisted QR payload from creation; contains request_id and contract only, no secret."
          },
          "status": {
            "type": "string",
            "enum": [
              "locked",
              "released",
              "refunded"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
...
      "RateLimited": {
        "description": "Rate limit exceeded for this route (per client IP).",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying.",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RateLimitError"
            }
          }
        }
      }
    }
  }
}